修改项目管理fcy字段为bigdecimal类型

+++ share-auth/share-auth-manage/share-auth-service/src/main/java/com/share/auth/model/entity/UemProject.java
@@ -18,6 +18,7 @@ import javax.persistence.Version;
 import java.lang.Integer;
 import java.lang.Long;
 import java.lang.String;
+import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -108,7 +109,7 @@ private static final long serialVersionUID=1;
 
     /**项目金额*/
     @Column(name = "fcy")
-    private Integer fcy;
+    private BigDecimal fcy;
 
     /**需求组员ID（逗号分隔）*/
     @Column(name = "gen_demand_users")
